% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFoldChange.R
\name{calculateFoldChange}
\alias{calculateFoldChange}
\title{calculateFoldChange}
\usage{
calculateFoldChange(df, type = "median", case, control)
}
\arguments{
\item{df}{Dataframe with a 'feature' column with feature names, a 'sample_id' column with sample ID's, a 'value' column with feature values for the samples, and a 'group' column splitting the samples into two groups (cases and controls).}

\item{type}{Calculate fold change from mean or median values? Default is 'median'.}

\item{case}{What are cases called in the 'group' column?}

\item{control}{What are controls called in the 'group' column?}
}
\value{
A dataframe with features and fold change/log2 fold change between cases and controls.
}
\description{
Calculates the fold change and log2 fold change between cases and controls.
}
