% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCarbonsDoubleBonds.R
\name{countCarbonsDoublebonds}
\alias{countCarbonsDoublebonds}
\title{countCarbonsDoubleBonds}
\usage{
countCarbonsDoublebonds(df)
}
\arguments{
\item{df}{Dataframe with lipid species names in the first column. Nomenclature must be xx:x for carbons:double_bonds.}
}
\value{
The dataframe with two additional columns 'carbons' and 'double_bonds'.
}
\description{
Count the number of carbons and double bonds in lipid species.
}
\examples{
data <- tibble(
'feature' = c('AC(6:0)', 'AC(8:1)', 'AC(10:2)',
             'Cer(d18:0/16:0)', 'Cer(d18:0/24:1)', 'Cer(d18:1/24:1)',
             'DAG(18:0/18:1)', 'DAG(16:1/18:1)', 'DAG(18:2/18:3)'))
counted <- data \%>\% countCarbonsDoublebonds()

}
