% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPCA.R
\name{createPCA}
\alias{createPCA}
\title{createPCA}
\usage{
createPCA(
  df,
  feature_annotations = "no",
  principal_components = c(1, 2),
  plots = "scores",
  addLabels = FALSE,
  cutoff,
  features = NULL
)
}
\arguments{
\item{df}{A tidy dataframe with at least sample_id, group, feature and value columns.}

\item{feature_annotations}{Does dataframe have a feature_anno column that can be used to color the features? 'yes' or 'no', default 'no'.}

\item{principal_components}{Principal components to be used as x and y axis, default is 1 and 2.}

\item{plots}{Which plots should be made? 'scores', 'loadings' or 'both'.}

\item{addLabels}{Add labels to features >cutoff or <-cutoff. TRUE or FALSE.}

\item{cutoff}{If addLabels is TRUE, which cutoff should be used?}
}
\value{
A PCA loadings plot.
}
\description{
Creates a PCA loadings plot using ggplot.
}
