% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideCols.R
\name{divideCols}
\alias{divideCols}
\title{divideCols}
\usage{
divideCols(df, col1, col2)
}
\arguments{
\item{df}{Dataframe with the columns.}

\item{col1}{Numerator.}

\item{col2}{Denominatr.}
}
\value{
A vector with the results.
}
\description{
Divides values in one column with values in another column.
}
