% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doOPLS.R
\name{doOPLS}
\alias{doOPLS}
\title{doOPLS}
\usage{
doOPLS(df)
}
\arguments{
\item{df}{A tidy dataframe with at least sample_id, group, feature and value column.}
}
\value{
A (r)opls OPLS object. Autoscaling is used on feature values.
}
\description{
Run ropls.
}
