% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCthreshold.R
\name{FCthreshold}
\alias{FCthreshold}
\title{FCthreshold}
\usage{
FCthreshold(
  df,
  p_threshold = 0.05,
  FC_threshold = 1.5,
  feature_annotations = "yes"
)
}
\arguments{
\item{df}{Dataframe with a p.adj column from rstatix and a log2(FC) column with the log2 fold change values.}

\item{p_threshold}{Adjusted p.value threshold. Default 0.05.}

\item{FC_threshold}{log2 fold_change threshold, default 1.5.}

\item{feature_annotations}{Does df have a 'feature_anno' column that can be used to color and annotate the features? If yes, this will be used. If default no features will be colored by their status as either up/downregulated or below threshold.}
}
\value{
The dataframe with a threshold column.
}
\description{
Add a 'threshold' column to change the shape of datapoints in the volcano-plot.
}
