% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_manuals.R
\name{featureAnnotationColors}
\alias{featureAnnotationColors}
\title{featureAnnotationColors}
\usage{
featureAnnotationColors(df)
}
\arguments{
\item{df}{The dataframe with data to be plotted.}
}
\description{
Create a color manual for the feature annotation factor to be used in plots.
}
