% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLS.R
\name{getOPLScorrcov}
\alias{getOPLScorrcov}
\title{getOPLScorrcov}
\usage{
getOPLScorrcov(opls_object, org_df)
}
\arguments{
\item{opls_object}{A ropls object.}

\item{ord_df}{The original df used to doOPLS.}
}
\value{
A dataframe with the correlations and covariance between opls principal score and the different features.
}
\description{
Get the covariance and correlations between the scores and the features used to calculate them.
}
