% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLS.R
\name{getOPLSloadings}
\alias{getOPLSloadings}
\title{getOPLSloadings}
\usage{
getOPLSloadings(opls_object)
}
\arguments{
\item{opls_object}{A ropls object.}
}
\value{
A dataframe with the principal and orthogonal loadings for each feature
}
\description{
Extract feature loadings from a (r)opls object.
}
