% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLS.R
\name{getOPLSorthovip}
\alias{getOPLSorthovip}
\title{getOPLSorthovip}
\usage{
getOPLSorthovip(opls_object)
}
\arguments{
\item{opls_object}{A ropls object.}
}
\value{
A dataframe with the feature and their orthogonal VIP score.
}
\description{
Extract feature and their orthogonal VIP scores from a (r)opls object.
}
