% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLS.R
\name{getOPLSpermutations}
\alias{getOPLSpermutations}
\title{getOPLSpermutations}
\usage{
getOPLSpermutations(opls_object)
}
\arguments{
\item{opls_object}{A ropls object.}
}
\value{
A dataframe with the R2Y, Q2Y values and simulation number.
}
\description{
Extract permutation results from a (r)opls object.
}
