% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLS.R
\name{getOPLSscores}
\alias{getOPLSscores}
\title{getOPLSscores}
\usage{
getOPLSscores(opls_object)
}
\arguments{
\item{opls_object}{A ropls object.}
}
\value{
A dataframe with the principal and orthogonal scores for each sample.
}
\description{
Extract the principal and orthogonal scores from a (r)opls object.
}
