% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLS.R
\name{getOPLSsummary}
\alias{getOPLSsummary}
\title{getOPLSsummary}
\usage{
getOPLSsummary(opls_object)
}
\arguments{
\item{opls_object}{A ropls object.}
}
\value{
A dataframe with the summary statistics provided by (r)opls.
}
\description{
Extract summary statistics from a (r)opls object.
}
