% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLS.R
\name{getOPLSvip}
\alias{getOPLSvip}
\title{getOPLSvip}
\usage{
getOPLSvip(opls_object)
}
\arguments{
\item{opls_object}{A ropls object.}
}
\value{
A dataframe with the feature and their principal VIP scores from a (r)opls object.
}
\description{
Extract feature principal component VIPs from a (r)opls object.
}
