% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOPLS.R
\name{plotAllOPLS}
\alias{plotAllOPLS}
\title{plotAllOPLS}
\usage{
plotAllOPLS(
  opls_object,
  org_df,
  metadata = TRUE,
  addLabels = FALSE,
  cov_cutoff = NULL,
  corr_cutoff = NULL,
  VIP_cutoff = NULL
)
}
\arguments{
\item{opls_object}{A ropls object.}

\item{addLabels}{If TRUE, adds labels to features above threshold values, if false no features have labels.}

\item{cov_cutoff}{Any feature with a covariance score below this is colored gray.}

\item{corr_cutoff}{Any feature with a correlation score below this is colored gray}

\item{VIP_cutoff}{Any feature with a principal VIP score below this is colored gray.}

\item{ord_df}{The original df used to doOPLS.}

\item{meta}{Should setMetaData be used to add metadata to the plots? Default TRUE.}
}
\value{
A white plot element with the summary statistics of ropls opls model.
}
\description{
Plots an OPLS score plot, loading plot and s-plot.
}
