% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOPLS.R
\name{plotOPLSloadings}
\alias{plotOPLSloadings}
\title{plotOPLSloadings}
\usage{
plotOPLSloadings(df, cutoff = NULL, features = NULL)
}
\arguments{
\item{df}{A 'getOPLSloadings' result dataframe, with p1 and o1 in the second and third column respectively, and a feature_anno column to color features by.}

\item{cutoff}{Numerical value used to label all features with a loading >cutoff or <-cutoff. Default NULL.}

\item{features}{Character string of features to be labelled, default NULL.}
}
\value{
A OPLS loading plot.
}
\description{
Plot the OPLS principal component and first orthogonal component loadings in ggplot2.
}
