% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOPLS.R
\name{plotOPLSpermutations}
\alias{plotOPLSpermutations}
\title{plotOPLSpermutations}
\usage{
plotOPLSpermutations(df)
}
\arguments{
\item{df}{A 'getOPLSpermutations' result dataframe.}
}
\value{
A permutation plot.
}
\description{
Plot the results from the permutation test run in (r)opls.
}
