% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOPLS.R
\name{plotOPLSscores}
\alias{plotOPLSscores}
\title{plotOPLSscores}
\usage{
plotOPLSscores(df)
}
\arguments{
\item{df}{A 'getOPLSscores' result dataframe.}
}
\value{
A OPLS scores plot, colored by group status.
}
\description{
Plot the OPLS principal component and first orthogonal component scores in ggplot2.
}
