% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOPLS.R
\name{plotOPLSsplot}
\alias{plotOPLSsplot}
\title{plotOPLSsplot}
\usage{
plotOPLSsplot(
  df,
  cov_cutoff = NULL,
  corr_cutoff = NULL,
  VIP_cutoff = NULL,
  cutoff_labels = FALSE,
  features = NULL
)
}
\arguments{
\item{df}{A 'getOPLScorrcov' result dataframe with a feature_anno column to color features.}

\item{cov_cutoff}{Any feature with a covariance score below this is colored gray.}

\item{corr_cutoff}{Any feature with a correlation score below this is colored gray}

\item{VIP_cutoff}{Any feature with a principal VIP score below this is colored gray.}

\item{cutoff_labels}{If TRUE, adds labels to features above threshold values, if false no features have labels.}

\item{features}{Character string of features to be labelled, default NULL.}
}
\value{
A white plot element with the summary statistics of ropls opls model.
}
\description{
Creates an s-plot for the (r)opls model.
}
