% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOPLS.R
\name{plotOPLSsummary}
\alias{plotOPLSsummary}
\title{plotOPLSscores}
\usage{
plotOPLSsummary(df)
}
\arguments{
\item{df}{A 'getOPLSsummary' result dataframe.}
}
\value{
A white plot element with the summary statistics of ropls opls model.
}
\description{
Create white plot element with the summary statistics of ropls opls model to be added to other OPLS plots.
}
