% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCAloadings.R
\name{plotPCAloadings}
\alias{plotPCAloadings}
\title{plotPCAloadings}
\usage{
plotPCAloadings(df, cutoff = NULL)
}
\arguments{
\item{df}{A dataframe with the principal components to be plotted in the first two columns.}

\item{cutoff}{If addLabels is TRUE, which cutoff should be used?}

\item{addLabels}{Add labels to features >cutoff or <-cutoff. TRUE or FALSE.}
}
\value{
A PCA loadings plot.
}
\description{
Creates a PCA loadings plot using ggplot.
}
