% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMetaData.R
\name{setMetaData}
\alias{setMetaData}
\title{setMetaData}
\usage{
setMetaData(df, sample_col = "group", feature_col = "feature_anno")
}
\arguments{
\item{df}{The dataframe to collect metadata from.}

\item{sample_col}{The columns to select as metadata for samples, default is 'group'. Set as FALSE if not in use.}

\item{feature_col}{The columns to select as metadata for features, default is 'feature_anno'.Set as FALSE if not in use.}
}
\value{
A list with sample and/or feature metadata.
}
\description{
Create a list of metadata for samples and features.
}
