% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study2018.R
\docType{data}
\name{study2018}
\alias{study2018}
\title{Metabolomics data from ME/CFS patients and healthy controls.}
\format{
A tibble with  40239 rows and 6 variables.
\describe{
  \item feature. Name of metabolites.
  \item feature_anno. What type of feature is it?
  \item SUB PATHWAY. What pathway does this feature belong to?
  \item sample_id. Individual sample id's forr all participants in the study.
  \item value. Concentration value of the feature for the sample.
  \item group. Shows if samples are in the ME/CFS or control group}
}
\usage{
data(study2018)
}
\description{
Taken from Prospective Biomarkers from Plasma Metabolomics of Myalgic Encephalomyelitis/Chronic Fatigue Syndrome Implicate Redox Imbalance in Disease Symptomatology. Metabolites. 2018 Dec 6;8(4):90. doi: 10.3390/metabo8040090.
}
\keyword{datasets}
